refactoring code

+++ controllers/product_controller.py
@@ -8,72 +8,59 @@ from models.product import ProductModel, ProductLocationModel
 
 from models.stock_location import StockLocationModel
 
-from api.stock import get_stockable_product
+from api.stock import ApiConnection
 
 
 class ProductController():
-    def __init__(self, view_master= None, columns= None, db_name= None):
+    def __init__(self, view_master= None, db=None, columns= None, db_name= None):
 
         self.columns= columns
         self.db_name= db_name
-        self.model= ProductModel()
-        self.stock_location_model = StockLocationModel()
-        self.product_location_model = ProductLocationModel()
+        self.api_connection = ApiConnection()
+        self.model= ProductModel(db=db)
+        self.stock_location_model = StockLocationModel(db=db)
+        self.product_location_model = ProductLocationModel(db=db)
         self.view_master= view_master
         self.location_values_id = {}
         self.product_frame = self.get_view()
-        
-
-    def create(self, data):
-        # print("create button clicked")
-        # print(data)
-        # id = self.model.create_query(data)
-        # keys = list(data.keys())
-        # for key in keys:
-        #     data[key.casefold()]= data.get(key)
-        #     data.pop(key)
-
-        # data['id'] = id
-        # print(data)
-        # color, bg_color = ('green','#e0e0e0') if len(self.user_frame.user_list.frame_list)%2 == 1 else ('blue violet','#C0C0C0')
-
-        # self.stock_location_frame.stock_location_list.create_elements(data, color, bg_color)
-        pass
 
 
     def refresh(self):
-        print('click refresh')
-        api_products = get_stockable_product()
+        api_products = self.api_connection.get_stockable_product()
 
-        db_ids = self.model.select_query(columns=['ODOO_ID'])
-        db_ids_dict = {}
-        for id in db_ids:
-            db_ids_dict[str(id[0])]= 1
-        #db_ids = list(map(lambda x: db_ids_dict[str(x[0])]= 1, db_ids))
-        
-        self.product_location_model.delete_all()
-        # string have all
-        product_ids = ''
+        if api_products:
+            db_ids = self.model.select_query(columns=['ODOO_ID'])
+            #self.model.delete_all_query()
+            db_ids_dict = {}
+            for id in db_ids:
+                db_ids_dict[str(id[0])]= 1
+            #db_ids = list(map(lambda x: db_ids_dict[str(x[0])]= 1, db_ids))
+            
+            print(db_ids)
+            self.product_location_model.delete_all()
+            # string have all
+            product_ids = ''
 
-        
-        for rec in api_products:
-            product_ids += str(rec['id'])+','
-            data = {}
-            print(rec['id'])
-            if not db_ids_dict.get(str(rec['id'])):
-                data = {'ODOO_ID': rec['id'], 'NAME': rec['name']}
-                product_id = self.model.create_query(data)
+            for rec in api_products:
                 
-                if rec.get('location_id'):
-                    for location in rec.get('location_id'):
-                        location_id = self.stock_location_model.select_query(columns=['ID'], conditions={'ODOO_ID': location})
-                        data = {'STOCK_LOCATION_ID': location_id[0][0], 'PRODUCT_ID':product_id}
-                        self.product_location_model.create_query(data)
-                        print('add product')
-        product_ids = product_ids[:-1]
-
-        self.model.delete_query(product_ids)
-        self.initialize_view(self.product_frame)
+                data = {}
+                if not db_ids_dict.get(str(rec['id'])):
+                    data = {'ODOO_ID': rec['id'], 'NAME': rec['name']}
+                    product_id = self.model.create_query(data)
+                    print(rec)
+                    if rec.get('location_id'):
+                        for location in rec.get('location_id'):
+                            location_id = self.stock_location_model.select_query(columns=['ID'], conditions={'ODOO_ID': location})
+                            data = {'STOCK_LOCATION_ID': location_id[0][0], 'PRODUCT_ID':product_id}
+                            self.product_location_model.create_query(data)
+                            print('add product')
+                else: product_ids += str(rec['id'])+','
+            product_ids = product_ids[:-1]
+            print('\n\n\n\n\n')
+            print(product_ids)
+            # self.model.delete_query(product_ids)
+            self.initialize_view(self.product_frame)
+    
 
 
         # # get stock location from odoo api
@@ -129,6 +116,7 @@ class ProductController():
 
         location_id = self.location_values_id.get(product_view.location.get())
         data = self.model.get_data(location_id)
+        print(location_id)
         for idx,instance in enumerate(data):
             instance_dict = {'id': instance[0],'ODOO_ID': instance[1], 'PRODUCT':instance[2]}
